ai-design-agent/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langgraph==0.1.5
langchain-google-genai==1.0.5
langchain-core==0.2.5
playwright==1.44.0
pydantic==2.7.1
python-dotenv==1.0.1
pillow==10.3.0
httpx==0.27.0
websockets==12.0
python-multipart==0.0.9